rename the timeline component
