+++ react/src/views/Transfers/ShipmentView/components/ShipmentTabs.tsx
@@ -1,5 +1,5 @@
 import { TabList, TabPanels, Tabs, TabPanel, Tab, Center } from "@chakra-ui/react";
-import ShipmentHistory, { IGroupedRecordEntry } from "components/Timeline/TimelineContainer";
+import ShipmentHistory, { IGroupedRecordEntry } from "components/Timeline/Timeline";
 import _ from "lodash";
 import { Box, ShipmentDetail, ShipmentState, User } from "types/generated/graphql";
 import ShipmentContent, { IShipmentContent } from "./ShipmentContent";
